[PC-538] ImageResizer Mock -> Fake로 변경
